Some atributes in SQL and in Dtos. Need to convert reader to stream. FK checks, Geometry fix.

+++ app/Ruian_Puller/src/main/java/cca/ruian_puller/download/dto/StavebniObjektDto.java
@@ -29,7 +29,7 @@ public class StavebniObjektDto {
     private LocalDateTime platido;
     private Long idtransakce;
     private Long globalniidnavrhuzmeny;
-    private Integer isknbudovaid;
+    private Long isknbudovaid;
     private LocalDateTime dokonceni;
     private Integer druhkonstrukcekod;
     private Integer obestavenyprostor;
